+++ src/app/code/community/Afterpay/Afterpay/Block/Onetouch.php
@@ -1,5 +1,9 @@
 <?php
 
+use Afterpay_Afterpay_Model_Method_Base as Afterpay_Base;
+use Afterpay_Afterpay_Model_System_Config_Source_CartMode as CartMode;
+use Afterpay_Afterpay_Model_System_Config_Source_ApiMode as ApiMode;
+
 class Afterpay_Afterpay_Block_Onetouch extends Mage_Core_Block_Template
 {
     /**
@@ -9,15 +13,12 @@ class Afterpay_Afterpay_Block_Onetouch extends Mage_Core_Block_Template
      */
     protected function _toHtml()
     {
-        $total = $this->getTotalAmount();
-        if  ( Mage::getStoreConfigFlag('afterpay/payovertime_cart/show_onetouch')
-                && Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Afterpay_Model_Method_Base::API_ENABLED_FIELD)
-                && Mage::helper('afterpay/checkout')->noConflict()
-                && Mage::getModel('afterpay/method_payovertime')->canUseForCheckoutSession()
-                && $total > 0
-                && $total >= Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Afterpay_Model_Method_Base::API_MIN_ORDER_TOTAL_FIELD)
-                && $total <= Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Afterpay_Model_Method_Base::API_MAX_ORDER_TOTAL_FIELD)
-            ) {
+        if ( Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Base::API_ENABLED_FIELD) &&
+            Mage::getStoreConfig('afterpay/payovertime_cart/show_onetouch') != CartMode::NO &&
+            Mage::helper('afterpay/checkout')->noConflict() &&
+            Mage::getModel('afterpay/method_payovertime')->canUseForCheckoutSession() &&
+            $this->isQuoteWithinLimits()
+        ) {
             return parent::_toHtml();
         } else {
             return '';
@@ -42,5 +43,46 @@ class Afterpay_Afterpay_Block_Onetouch extends Mage_Core_Block_Template
     public function getTotalAmount()
     {
         return Mage::helper('afterpay')->calculateTotal();
+    }
+
+    public function isQuoteWithinLimits()
+    {
+        $total = $this->getTotalAmount();
+        $min = Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Base::API_MIN_ORDER_TOTAL_FIELD);
+        $max = Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Base::API_MAX_ORDER_TOTAL_FIELD);
+
+        return ($total > 0 && $total >= $min && $total <= $max);
+    }
+
+    public function isExpress()
+    {
+        return Mage::getStoreConfig('afterpay/payovertime_cart/show_onetouch') == CartMode::EXPRESS;
+    }
+
+    public function isShippingRequired()
+    {
+        $quote = Mage::getSingleton('checkout/session')->getQuote();
+        return !$quote->isVirtual();
+    }
+
+    public function getCountryCode()
+    {
+        $countryCode = '';
+        $currency = ApiMode::getCurrencyCode();
+
+        if (array_key_exists($currency, Afterpay_Base::CURRENCY_PROPERTIES)){
+            $countryCode = Afterpay_Base::CURRENCY_PROPERTIES[$currency]['jsCountry'];
+        }
+
+        return $countryCode;
+    }
+
+    public function getJsUrl()
+    {
+        $apiMode = Mage::getStoreConfig('payment/afterpaypayovertime/' . Afterpay_Base::API_MODE_CONFIG_FIELD);
+        $settings = ApiMode::getEnvironmentSettings($apiMode);
+        $key = urlencode(Mage::getStoreConfig('afterpay/payovertime_cart/express_key'));
+
+        return $settings[ApiMode::KEY_WEB_URL] . 'afterpay.js?merchant_key=' . $key;
     }
 }