+++ src/js/Afterpay/checkout/awesomecheckout.js
@@ -16,7 +16,7 @@
                  */
                 // Run ajax to process Afterpay payment using Protorype
                 var request = new Ajax.Request(
-                    window.Afterpay.saveUrl, // use Afterpay controller
+                    window.AfterpayM1.saveUrl, // use Afterpay controller
                     {
                         method: 'post',
                         onSuccess: function (transport) {
@@ -47,7 +47,7 @@
                                     });
                                 }
 
-                                switch (window.Afterpay.redirectMode) {
+                                switch (window.AfterpayM1.redirectMode) {
                                     case 'lightbox':
                                         AfterPay.display({
                                             token: response.token